refactor(cats-core): Introduce the CachingAgentDataTypes class
Most clients want the set of authoritative types. Some clients want to
know all the types. But literally everyone wants to filter and transform
Collection<AgentDataType> into some sort of Set<String>, so let's just
make that easier by providing a data object with a few different sets.

I also kinda added a third type for On-Demand. Once every caching agent
is updated to define this, we can strengthen the API contract of
CachingAgent and log warnings if any of these things are not true:
a) only things in authoritative and on-demand have attributes
b) only things in authoritative and informative have relationships
c) there are no types in the result other than those listed in
   authoritative, informative, and on-demand

Removing AgentDataType will be done in a future pull request.