fix(cats-core): Remove cached data that didn't have a CacheResult key
If a caching agent has some serverGroups stored, but in the next run it
finds no server groups, it may just not put the serverGroups key into
the CacheResult. Currently, that means those server groups will stick
around in the cache forever. They don't have any relationships pointing
to them, so they're just orphaned, but their attributes are still stored
forever.

Amazon doesn't have this problem because they always stick a
List<CacheData> under the serverGroups key, even if it's empty. This is
enough to tell putCacheResult() to store that empty list, overwriting
whatever was there before.

This commit makes it so that we treat a missing key the same as if there
was an empty list there. We could update all the other caching agents to
work like the AWS ones, but this seems like a more correct fix.

The orphaned data isn't the cause of spinnaker/spinnaker#4511, but it
dramatically exacerbates that issue.